Implement user authentication and authorization features

+++ src/services/notes/controller/note-controller.js
@@ -1,13 +1,19 @@
 import NoteRepositories from '../repositories/note-repositories.js';
 import { InvariantError, NotFoundError } from '../../../exceptions/index.js';
 import response from '../../../utils/response.js';
+import AuthorizationError from '../../../exceptions/authorization-error.js';
 
 export const createNote = async (req, res, next) => {
   const { title, body, tags } = req.validated;
+  const { id: owner } = req.user;
+
+  console.log('OWNER SAAT CREATE:', owner);
+
   const note = await NoteRepositories.createNote({
     title,
     body,
-    tags
+    tags,
+    owner
   });
 
   if (!note) {
@@ -20,17 +26,18 @@ export const createNote = async (req, res, next) => {
 };
 
 export const getNotes = async (req, res) => {
-  const notes = await NoteRepositories.getNotes();
-  return response(res, 200, 'Catatan sukses ditampilkan', { notes: notes });
+  const { id: owner } = req.user;
+  const notes = await NoteRepositories.getNotes(owner);
+  return response(res, 200, 'Catatan sukses ditampilkan', { notes });
 };
 
 export const getNoteById = async (req, res, next) => {
   const { id } = req.params;
-  const note = await NoteRepositories.getNoteById(id);
+  const { id: owner } = req.user;
 
-  if (!note) {
-    return next(new NotFoundError('Catatan tidak ditemukan'));
-  }
+  await NoteRepositories.verifyNoteOwner(id, owner);
+
+  const note = await NoteRepositories.getNoteById(id);
 
   return response(res, 200, 'Catatan sukses ditampilkan', {
     note,
@@ -45,6 +52,10 @@ export const editNoteById = async (req, res, next) => {
     tags
   } = req.validated;
 
+  const { id: owner } = req.user;
+
+  await NoteRepositories.verifyNoteOwner(id, owner);
+
   const note = await NoteRepositories.editNote({
     id,
     title,
@@ -56,16 +67,24 @@ export const editNoteById = async (req, res, next) => {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
 
-  return response(res, 200, 'Catatan berhasil diperbarui');
+  return response(res, 200, 'Catatan berhasil diperbarui', {
+    note
+  });
 };
 
 export const deleteNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  await NoteRepositories.verifyNoteOwner(id, owner);
+
   const deletedNote = await NoteRepositories.deleteNote(id);
 
   if (!deletedNote) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
 
-  return response(res, 200, 'Catatan berhasil dihapus');
+  return response(res, 200, 'Catatan berhasil dihapus', {
+    deletedNoteId: deletedNote,
+  });
 };
